requirements.txt
boto3
aws-cdk.core==1.124.0
aws_cdk.aws_ecs==1.124.0
aws_cdk.aws_eks==1.124.0
aws-cdk.aws-iam==1.124.0
aws-cdk.aws-autoscaling==1.124.0